remove the login page and made the google login in the landing page
