+++ frontend/src/App.jsx
@@ -2,7 +2,6 @@ import { useMemo, useState } from 'react';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import HomePage from '@/pages/HomePage';
 import LandingPage from '@/pages/LandingPage';
-import LoginPage from '@/pages/LoginPage';
 import ProfilePage from '@/pages/ProfilePage';
 import UserProfile from '@/pages/UserProfile';
 
@@ -44,18 +43,8 @@ function App() {
     <BrowserRouter>
       <Routes>
         <Route
-          path='/login'
-          element={
-            currentUser ? (
-              <Navigate to='/' replace />
-            ) : (
-              <LoginPage onLogin={authApi.login} />
-            )
-          }
-        />
-        <Route
           path='/landing'
-          element={currentUser ? <Navigate to='/' replace /> : <LandingPage />}
+          element={currentUser ? <Navigate to='/' replace /> : <LandingPage onLogin={authApi.login} />}
         />
         <Route
           path='/'
@@ -73,7 +62,7 @@ function App() {
             currentUser ? (
               <ProfilePage currentUser={currentUser} onUserUpdate={authApi.updateUser} />
             ) : (
-              <Navigate to='/login' replace />
+              <Navigate to='/landing' replace />
             )
           }
         />
@@ -83,7 +72,7 @@ function App() {
             currentUser ? (
               <UserProfile currentUser={currentUser} />
             ) : (
-              <Navigate to='/login' replace />
+              <Navigate to='/landing' replace />
             )
           }
         />
