+++ frontend/src/pages/LandingPage.jsx
@@ -1,11 +1,55 @@
+import { useEffect, useRef, useState } from 'react';
 import { motion } from 'motion/react';
-import { useNavigate } from 'react-router-dom';
 import {
   MapPin,
   Navigation,
   ShieldCheck,
   Star,
 } from 'lucide-react';
+
+const API_ROOT = (import.meta.env.VITE_API_BASE_URL || '/api').replace(/\/$/, '');
+const GOOGLE_CLIENT_ID = import.meta.env.VITE_GOOGLE_CLIENT_ID;
+
+function loadGoogleScript() {
+  return new Promise((resolve, reject) => {
+    const waitForGoogle = (attempt = 0) => {
+      if (window.google?.accounts?.id) {
+        resolve();
+        return;
+      }
+      if (attempt >= 40) {
+        reject(new Error('Google script loaded but is unavailable. Disable blockers and allow accounts.google.com.'));
+        return;
+      }
+      window.setTimeout(() => waitForGoogle(attempt + 1), 100);
+    };
+
+    if (window.google?.accounts?.id) {
+      resolve();
+      return;
+    }
+
+    const existing = document.querySelector('script[data-google-gsi="true"]');
+    if (existing) {
+      if (existing.dataset.loaded === 'true') {
+        waitForGoogle();
+        return;
+      }
+      existing.addEventListener('load', () => { existing.dataset.loaded = 'true'; waitForGoogle(); }, { once: true });
+      existing.addEventListener('error', () => reject(new Error('Failed to load Google Identity Services')), { once: true });
+      return;
+    }
+
+    const script = document.createElement('script');
+    script.src = 'https://accounts.google.com/gsi/client';
+    script.async = true;
+    script.defer = true;
+    script.dataset.googleGsi = 'true';
+    script.onload = () => { script.dataset.loaded = 'true'; waitForGoogle(); };
+    script.onerror = () => reject(new Error('Failed to load Google Identity Services'));
+    document.head.appendChild(script);
+  });
+}
 
 function GoogleIcon({ className }) {
   return (
@@ -26,9 +70,81 @@ function GithubIcon({ className }) {
   );
 }
 
-export default function LandingPage() {
-  const navigate = useNavigate();
-  const handleSignIn = () => navigate('/login');
+export default function LandingPage({ onLogin }) {
+  const googleReady = useRef(false);
+  const [error, setError] = useState('');
+  const [isLoading, setIsLoading] = useState(true);
+  const [signingIn, setSigningIn] = useState(false);
+
+  useEffect(() => {
+    let isMounted = true;
+
+    async function init() {
+      if (!GOOGLE_CLIENT_ID) {
+        if (isMounted) {
+          setError('Missing VITE_GOOGLE_CLIENT_ID in frontend/.env.');
+          setIsLoading(false);
+        }
+        return;
+      }
+
+      try {
+        await loadGoogleScript();
+        if (!isMounted) return;
+
+        window.google.accounts.id.initialize({
+          client_id: GOOGLE_CLIENT_ID,
+          callback: async (response) => {
+            try {
+              if (!response?.credential) throw new Error('Google did not return a credential token.');
+
+              const authResponse = await fetch(`${API_ROOT}/auth/google`, {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+                body: JSON.stringify({ credential: response.credential }),
+              });
+
+              if (!authResponse.ok) {
+                const payload = await authResponse.json().catch(() => ({}));
+                throw new Error(payload?.error || 'Google login failed');
+              }
+
+              const payload = await authResponse.json();
+              if (!payload?.user) throw new Error('Google login response is missing user data');
+
+              onLogin(payload.user);
+            } catch (err) {
+              setError(err instanceof Error ? err.message : 'Login failed');
+              setSigningIn(false);
+            }
+          },
+        });
+
+        googleReady.current = true;
+        if (isMounted) setIsLoading(false);
+      } catch (err) {
+        if (isMounted) {
+          setError(err instanceof Error ? err.message : 'Failed to initialize login');
+          setIsLoading(false);
+        }
+      }
+    }
+
+    init();
+    return () => { isMounted = false; };
+  }, [onLogin]);
+
+  const handleSignIn = () => {
+    if (!googleReady.current) return;
+    setSigningIn(true);
+    setError('');
+    window.google.accounts.id.prompt((notification) => {
+      if (notification.isNotDisplayed() || notification.isSkippedMoment()) {
+        setSigningIn(false);
+        setError('Sign-in prompt was blocked. Please disable your ad blocker or allow pop-ups for this site.');
+      }
+    });
+  };
 
   return (
     <div className="min-h-screen flex flex-col font-['Inter',sans-serif] bg-[#f8f9fa] text-[#001a48] antialiased">
@@ -40,13 +156,17 @@ export default function LandingPage() {
               HopShare
             </span>
           </div>
-          <button 
-            onClick={handleSignIn}
-            className="bg-[#002d72] text-white px-5 py-2 rounded-full text-sm font-bold flex items-center gap-2 hover:opacity-90 transition-all active:scale-95"
-          >
-            <GoogleIcon className="w-4 h-4" />
-            Sign in with Google
-          </button>
+          <div className="flex flex-col items-end gap-1">
+            <button
+              onClick={handleSignIn}
+              disabled={isLoading || signingIn}
+              className="bg-[#002d72] text-white px-5 py-2 rounded-full text-sm font-bold flex items-center gap-2 hover:opacity-90 transition-all active:scale-95 disabled:opacity-60 disabled:cursor-not-allowed"
+            >
+              <GoogleIcon className="w-4 h-4" />
+              {signingIn ? 'Signing in…' : 'Sign in with Google'}
+            </button>
+            {error && <p className="text-xs text-red-600 max-w-xs text-right">{error}</p>}
+          </div>
         </div>
       </nav>
 
@@ -54,8 +174,8 @@ export default function LandingPage() {
       <main className="relative flex-grow pt-16 flex flex-col">
         {/* Background Img Layer */}
         <div className="absolute inset-0 z-0 overflow-hidden">
-          <img 
-            src="https://www.tripsavvy.com/thmb/85wFz774F2_2HV5MrYCW7xA53MI=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/GettyImages-1050146084-1f58e25cdab445b996e5896ee6e4cbe0.jpg" 
+          <img
+            src="https://www.tripsavvy.com/thmb/85wFz774F2_2HV5MrYCW7xA53MI=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/GettyImages-1050146084-1f58e25cdab445b996e5896ee6e4cbe0.jpg"
             alt="Rideshare photo"
             className="w-full h-full object-cover opacity-40 grayscale"
             referrerPolicy="no-referrer"
@@ -67,7 +187,7 @@ export default function LandingPage() {
         <section className="relative z-10 min-h-screen flex items-center justify-center px-6 py-12">
           <div className="max-w-7xl w-full flex flex-col lg:flex-row items-center gap-16">
             {/* Left Content */}
-            <motion.div 
+            <motion.div
               initial={{ opacity: 0, y: 20 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.6 }}
@@ -79,11 +199,11 @@ export default function LandingPage() {
                   Made for JHU Students, by JHU Students
                 </span>
               </div>
-              
+
               <h1 className="text-3xl md:text-5xl font-extrabold tracking-tight text-[#001a48] leading-[1.1] font-['Manrope',sans-serif]">
                 Get to wherever you're going...with your fellow <span className="text-[#046398]">Blue Jays.</span>
               </h1>
-              
+
               <p className="text-lg md:text-xl text-[#2d2f3a] max-w-xl leading-relaxed mx-auto lg:mx-0">
                 HopShare is the ultimate rideshare coordination platform built for the JHU community. Ride with other students and save money!
               </p>
@@ -108,7 +228,7 @@ export default function LandingPage() {
                   <div>
                     <p className="font-bold text-[#001a48]">Drivers: offer a seat</p>
                     <p className="text-sm text-[#2d2f3a] leading-relaxed">
-                      Already driving somewhere? Post your route and time so riders heading your way can reach out and hop in. You can earn a bit of extra money for a trip you were already going to make, or offer a free ride! 
+                      Already driving somewhere? Post your route and time so riders heading your way can reach out and hop in. You can earn a bit of extra money for a trip you were already going to make, or offer a free ride!
                     </p>
                   </div>
                 </div>
@@ -129,7 +249,7 @@ export default function LandingPage() {
             </motion.div>
 
             {/* Right Content: Floating Card */}
-            <motion.div 
+            <motion.div
               initial={{ opacity: 0, scale: 0.9, rotate: 2 }}
               animate={{ opacity: 1, scale: 1, rotate: 0 }}
               transition={{ duration: 0.8, delay: 0.2 }}
@@ -168,7 +288,7 @@ export default function LandingPage() {
                 <div className="flex items-center gap-4 p-4 bg-[#81c4ff] rounded-2xl">
                   <div className="flex -space-x-3">
                     {[1, 2, 3].map((i) => (
-                      <img 
+                      <img
                         key={i}
                         src={`https://picsum.photos/seed/student${i}/100/100`}
                         alt="Student"
